Workaround: Windows doesn't support symbolic links
We need to make a copy of the package directories that replace all
symbolic links with actual files. We don't want to replace in-place
because that could screw up the image for future iterations (also, copy
is easier to do). There are better solutions to this and for supporting
linux access but neither is worth fixing when we'll be changing a bunch
for container lambda support