Features needed for retrievers: specify zip, freeze, constraints
Allow the user to specify a zip_file to save the bundle to
Allow the user to specify a freeze file to save to
Allow supplying constraints files (and properly track if they change).

Also fixes bugs in freezing when using no_secrets.

Sorry, would have been a pain to separate these into their own commits by the time I was done with them